+++ UserSettings/EditorUserSettings.asset
@@ -30,10 +30,10 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c7b15051e1f24293a66032e2c1336acf53a31f6fe
       flags: 0
     RecentlyUsedScenePath-8:
-      value: 22424703114646680e0b0227036c721104033025202d68252320092a
+      value: 22424703114646680e0b0227036c72111f19563f22213229
       flags: 0
     RecentlyUsedScenePath-9:
-      value: 22424703114646680e0b0227036c72111f19563f22213229
+      value: 22424703114646680e0b0227036c721104033025202d68252320092a
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
